Rename: 컴포넌트 jsx 파일 index.jsx로 이름 변경
Related to: #12
