+++ src/components/atoms/Input/Input.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Input from 'components/atoms/Input/Input';
+import Input from 'components/atoms/Input';
 
 export default {
   title: 'Atoms/Input',
