+++ src/components/atoms/Label/Label.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Label from 'components/atoms/Label/Label';
+import Label from 'components/atoms/Label';
 
 export default {
   title: 'Atoms/Label',
