+++ src/components/atoms/LogoImage/index.jsx
