+++ src/components/molecules/LogList/LogList.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LogList from 'components/molecules/LogList/LogList';
+import LogList from 'components/molecules/LogList';
 
 export default {
   title: 'Molecules/LogList',
