+++ src/components/molecules/LogList/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from 'react';
 import mockData from './LogListMockData';
-import ListItem from 'components/atoms/ListItem/ListItem';
+import ListItem from 'components/atoms/ListItem';
 import * as Styled from 'components/molecules/LogList/LogList.style';
 
 const LogList = () => {
