+++ src/components/molecules/Logo/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import LogoTitle from 'components/atoms/LogoTitle/LogoTitle';
-import LogoImage from 'components/atoms/LogoImage/LogoImage';
+import LogoTitle from 'components/atoms/LogoTitle';
+import LogoImage from 'components/atoms/LogoImage';
 import * as Styled from 'components/molecules/Logo/Logo.style';
 
 const LogoImageIsFlipped = {
