+++ src/components/molecules/MoneyBox/MoneyBox.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MoneyBox from 'components/molecules/MoneyBox/MoneyBox';
+import MoneyBox from 'components/molecules/MoneyBox';
 
 export default {
   title: 'Molecules/MoneyBox',
