+++ src/components/molecules/NavButton/NavButton.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavButton from 'components/molecules/NavButton/NavButton';
+import NavButton from 'components/molecules/NavButton';
 
 export default {
   title: 'Molecules/NavButton',
