+++ src/components/molecules/ProductBox/ProductBox.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductBox from 'components/molecules/ProductBox/ProductBox';
+import ProductBox from 'components/molecules/ProductBox';
 
 export default {
   title: 'Molecules/ProductBox',
