+++ src/components/organisms/VendingMachine/VendingMachine.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import VendingMachine from 'components/organisms/VendingMachine/VendingMachine';
+import VendingMachine from 'components/organisms/VendingMachine';
 
 export default {
   title: 'Organisms/VendingMachine',
