+++ src/components/organisms/VendingMachine/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Logo from 'components/molecules/Logo/Logo';
-import ProductBox from 'components/molecules/ProductBox/ProductBox';
+import Logo from 'components/molecules/Logo';
+import ProductBox from 'components/molecules/ProductBox';
 import mockData from 'components/organisms/VendingMachine/ProductMockData';
 import { StyledVM, ProductWrapper } from 'components/organisms/VendingMachine/VendingMachine.style';
