+++ src/components/organisms/VendingMachineController/index.jsx
@@ -1,8 +1,8 @@
-import Button from 'components/atoms/Button/Button';
-import Label from 'components/atoms/Label/Label';
-import LogList from 'components/molecules/LogList/LogList';
-import NavButton from 'components/molecules/NavButton/NavButton';
-import UserInputBox from 'components/molecules/UserInputBox/UserInputBox';
+import Button from 'components/atoms/Button';
+import Label from 'components/atoms/Label';
+import LogList from 'components/molecules/LogList';
+import NavButton from 'components/molecules/NavButton';
+import UserInputBox from 'components/molecules/UserInputBox';
 import * as Styled from 'components/organisms/VendingMachineController/VendingMachineController.style';
 
 const VMController = () => {
