+++ src/pages/VendingMachine/index.jsx
@@ -1,6 +1,6 @@
 import * as Styled from 'pages/VendingMachine/VMPage.style';
-import VendingMachine from 'components/organisms/VendingMachine/VendingMachine';
-import VendingMachineController from 'components/organisms/VendingMachineController/VendingMachineController';
+import VendingMachine from 'components/organisms/VendingMachine';
+import VendingMachineController from 'components/organisms/VendingMachineController';
 
 const VMPage = () => {
   return (
