+++ src/pages/Wallet/index.jsx
@@ -1,6 +1,6 @@
-import Wallet from 'components/organisms/Wallet/Wallet';
+import Wallet from 'components/organisms/Wallet';
 import * as Styled from 'pages/Wallet/WalletPage.style';
-import VendingMachineController from 'components/organisms/VendingMachineController/VendingMachineController';
+import VendingMachineController from 'components/organisms/VendingMachineController';
 
 const WalletPage = () => {
   return (
